# api-docs.yaml
openapi: 3.1.0
info:
  title: GBCC Website API
  description: |
    REST API интернет-магазина / каталога GBCC: аутентификация (JWT в HttpOnly cookies),
    аккаунты, каталог товаров и файлы, новости, акции (скидки), заказы, обращения в поддержку.

    **Авторизация в Swagger UI:** схема `accessTokenCookie` — cookie `GBCC_ACCESS_TOKEN`
    (после успешного `POST /auth/login` в том же браузере cookie подставляется автоматически;
    при тестах из Postman — передайте cookie вручную).

    **Гостевой токен поддержки:** схема `supportGuestToken` — заголовок `X-Support-Token`
    (значение из `guestAccessToken` после создания обращения без входа).
  version: 0.2.0
servers:
- url: http://localhost:8080
  description: Local Dev Server
...
paths:
...
  /support/conversations:
    get:
      tags:
      - Support
      summary: Список обращений (администраторы)
      description: "Требует JWT. В сервисе допускаются только ADMIN и OWNER. Фильт\
        ры: status, query (тема или email гостя)."
      operationId: searchForAdmin
      parameters:
      - name: status
        in: query
        required: false
        schema:
          type: string
          enum:
          - OPEN
          - CLOSED
      - name: query
        in: query
        required: false
        schema:
          type: string
      responses:
        "200":
          description: Страница
          content:
            application/json: {}
        "401":
          description: Нет JWT
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "403":
          description: Не менеджер
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
      security:
      - accessTokenCookie: []
    post:
      tags:
      - Support
      summary: Создать обращение
      description: |
        Без JWT: обязательны message, guestName, guestEmail (валидный email).
        С JWT и ролью CUSTOMER: только message (и опционально subject); гостевой токен не выдаётся.
        Администраторам создавать диалог через этот метод нельзя — отвечают в существующем.
      operationId: create
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSupportConversationRequestDto"
        required: true
      responses:
        "201":
          description: Создано; для гостя в теле guestAccessToken
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CreateSupportConversationResponseDto"
        "400":
          description: Некорректные данные
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "403":
          description: Попытка создать от имени администратора
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
  /support/conversations/{conversationId}/messages:
    post:
      tags:
      - Support
      summary: Добавить сообщение
      description: |
        Приоритет: если валиден X-Support-Token для этого диалога — сообщение от гостя (даже при наличии JWT).
        Иначе: ADMIN/OWNER — ответ администратора; CUSTOMER — только владелец диалога.
        Закрытый диалог — ошибка 400.
      operationId: addMessage
      parameters:
      - name: conversationId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: X-Support-Token
        in: header
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddSupportMessageRequestDto"
        required: true
      responses:
        "200":
          description: Обновлённый диалог
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SupportConversationDetailResponseDto"
        "400":
          description: Диалог закрыт или неверные данные
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "403":
          description: Нет прав
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "404":
          description: Диалог не найден
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
...
  /cooperation-requests:
    get:
      tags:
      - Cooperation requests
      summary: Список заявок (администраторы)
      description: "JWT. Фильтры: onlyUnassigned, query (поиск по имени, email, те\
        лефону, комментарию, типу сотрудничества)."
      operationId: searchForAdmin_1
      parameters:
      - name: onlyUnassigned
        in: query
        required: false
        schema:
          type: boolean
      - name: query
        in: query
        required: false
        schema:
          type: string
      responses:
        "200":
          description: Страница заявок
          content:
            application/json: {}
        "401":
          description: Нет JWT
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "403":
          description: Не менеджер
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
      security:
      - accessTokenCookie: []
    post:
      tags:
      - Cooperation requests
      summary: Отправить заявку
      description: Публично. Требуется согласие на обработку персональных данных.
      operationId: create_5
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCooperationRequestDto"
        required: true
      responses:
        "201":
          description: Создана
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CooperationRequestResponseDto"
        "400":
          description: Валидация или файл не найден
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
...
  /contact-requests:
    get:
      tags:
      - Contact requests
      summary: Список заявок (администраторы)
      description: "JWT. Фильтры: onlyUnassigned (только без назначенного), query\
        \ (поиск по имени, email, телефону, тексту сообщения)."
      operationId: searchForAdmin_2
      parameters:
      - name: onlyUnassigned
        in: query
        required: false
        schema:
          type: boolean
      - name: query
        in: query
        required: false
        schema:
          type: string
      responses:
        "200":
          description: Страница заявок
          content:
            application/json: {}
        "401":
          description: Нет JWT
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "403":
          description: Не менеджер
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
      security:
      - accessTokenCookie: []
    post:
      tags:
      - Contact requests
      summary: Отправить заявку
      description: Публично. Требуется согласие на обработку персональных данных.
      operationId: create_6
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContactRequestDto"
        required: true
      responses:
        "201":
          description: Создана
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ContactRequestResponseDto"
        "400":
          description: Валидация
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
...
  /support/conversations/{conversationId}/status:
    patch:
      tags:
      - Support
      summary: Изменить статус обращения
      description: JWT. Только ADMIN/OWNER (например CLOSED).
      operationId: updateStatus
      parameters:
      - name: conversationId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSupportConversationStatusRequestDto"
        required: true
      responses:
        "200":
          description: Обновлено
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SupportConversationDetailResponseDto"
        "401":
          description: Нет JWT
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "403":
          description: Не менеджер
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "404":
          description: Не найдено
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
      security:
      - accessTokenCookie: []
  /order/{orderId}/status:
    patch:
      tags:
      - Orders
      summary: Изменить статус заказа
      description: |
        JWT. Только ADMIN/OWNER. Переходы статусов валидируются на сервере.
        Опционально: estimatedDeliveryAt / estimatedDeliveryEnd (окно доставки), receiptUrl (электронный чек), comment.
      operationId: updateStatus_1
      parameters:
      - name: orderId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusRequestDto"
        required: true
      responses:
        "200":
          description: Обновлено
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OrderResponseDto"
        "400":
          description: Недопустимый переход статуса
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "401":
          description: Нет JWT
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "403":
          description: Не менеджер
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "404":
          description: Заказ не найден
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
      security:
      - accessTokenCookie: []
...
  /support/conversations/{conversationId}:
    get:
      tags:
      - Support
      summary: Получить переписку по id
      description: |
        Доступ: JWT ADMIN/OWNER (любой диалог); JWT CUSTOMER (свой диалог); гость — заголовок X-Support-Token (схема supportGuestToken в Authorize).
      operationId: get
      parameters:
      - name: conversationId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: X-Support-Token
        in: header
        description: Токен гостя (если обращение без аккаунта)
        required: false
        schema:
          type: string
      responses:
        "200":
          description: Диалог с сообщениями
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SupportConversationDetailResponseDto"
        "403":
          description: Нет доступа
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "404":
          description: Не найдено
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
  /support/conversations/my:
    get:
      tags:
      - Support
      summary: Мои обращения (клиент)
      description: JWT. Только роль CUSTOMER. Пагинация Spring.
      operationId: listMine
      responses:
        "200":
          description: Страница обращений
          content:
            application/json: {}
        "401":
          description: Нет JWT
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "403":
          description: Не клиент
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
      security:
      - accessTokenCookie: []
...
components:
  schemas:
...
    ApiErrorField:
      type: object
      properties:
        field:
          type: string
        message:
          type: string
    ApiErrorResponse:
      type: object
      description: Единый формат ошибки API (см. GlobalExceptionHandler)
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          format: int32
        error:
          type: string
        message:
          type: string
        path:
          type: string
        fields:
          type: array
          items:
            $ref: "#/components/schemas/ApiErrorField"
...
    CreateSupportConversationRequestDto:
      type: object
      properties:
        subject:
          type: string
          maxLength: 500
          minLength: 0
        message:
          type: string
          maxLength: 16000
          minLength: 0
        guestName:
          type: string
          maxLength: 255
          minLength: 0
        guestEmail:
          type: string
          format: email
          maxLength: 255
          minLength: 0
      required:
      - message
    CreateSupportConversationResponseDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        guestAccessToken:
          type: string
        subject:
          type: string
        status:
          type: string
          enum:
          - OPEN
          - CLOSED
        createdAt:
          type: string
          format: date-time
    AddSupportMessageRequestDto:
      type: object
      properties:
        message:
          type: string
          maxLength: 16000
          minLength: 0
      required:
      - message
    SupportConversationDetailResponseDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        subject:
          type: string
        status:
          type: string
          enum:
          - OPEN
          - CLOSED
        guestName:
          type: string
        guestEmail:
          type: string
        customerId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        messages:
          type: array
          items:
            $ref: "#/components/schemas/SupportMessageResponseDto"
    SupportMessageResponseDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        authorType:
          type: string
          enum:
          - GUEST
          - CUSTOMER
          - ADMIN
        authorAccountId:
          type: string
          format: uuid
        body:
          type: string
        createdAt:
          type: string
          format: date-time
...
    OrderItemResponseDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        productId:
          type: string
          format: uuid
        productBrand:
          type: string
        productClassName:
          type: string
        productSeriesName:
          type: string
        productTypeName:
          type: string
        quantity:
          type: integer
          format: int32
        unitPrice:
          type: number
        unitDiscountPercent:
          type: number
        unitDiscountedPrice:
          type: number
        lineTotalPrice:
          type: number
        lineTotalDiscountedPrice:
          type: number
    OrderResponseDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        displayNumber:
          type: integer
          format: int64
        customerId:
          type: string
          format: uuid
        status:
          type: string
          enum:
          - CREATED
          - PROCESSING
          - PACKED
          - SHIPPED
          - DELIVERED
          - CANCELLED
        statusLabel:
          type: string
        deliveryAddress:
          type: string
        customerComment:
          type: string
        paymentMethod:
          type: string
          enum:
          - CARD_ONLINE
          - CARD_OR_ON_RECEIPT
        paymentMethodLabel:
          type: string
        receiptUrl:
          type: string
        estimatedDeliveryAt:
          type: string
          format: date-time
        estimatedDeliveryEnd:
          type: string
          format: date-time
        totalPrice:
          type: number
        totalDiscountedPrice:
          type: number
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemResponseDto"
        history:
          type: array
          items:
            $ref: "#/components/schemas/OrderStatusHistoryResponseDto"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    OrderStatusHistoryResponseDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        fromStatus:
          type: string
          enum:
          - CREATED
          - PROCESSING
          - PACKED
          - SHIPPED
          - DELIVERED
          - CANCELLED
        toStatus:
          type: string
          enum:
          - CREATED
          - PROCESSING
          - PACKED
          - SHIPPED
          - DELIVERED
          - CANCELLED
        changedByAccountId:
          type: string
          format: uuid
        comment:
          type: string
        estimatedDeliveryAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
...
    CreateCooperationRequestDto:
      type: object
      properties:
        name:
          type: string
          maxLength: 255
          minLength: 0
        phone:
          type: string
          maxLength: 64
          minLength: 0
        email:
          type: string
          format: email
          maxLength: 255
          minLength: 0
        cooperationType:
          type: string
          maxLength: 128
          minLength: 0
        comment:
          type: string
        attachmentFileId:
          type: string
          format: uuid
        consentProcessing:
          type: boolean
      required:
      - email
      - name
      - phone
    CooperationRequestResponseDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        cooperationType:
          type: string
        comment:
          type: string
        attachmentFileId:
          type: string
          format: uuid
        consentProcessing:
          type: boolean
        assignedToAccountId:
          type: string
          format: uuid
        assignedToName:
          type: string
        assignedToEmail:
          type: string
        assignedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateContactRequestDto:
      type: object
      properties:
        name:
          type: string
          maxLength: 255
          minLength: 0
        email:
          type: string
          format: email
          maxLength: 255
          minLength: 0
        phone:
          type: string
          maxLength: 64
          minLength: 0
        message:
          type: string
          minLength: 1
        consentProcessing:
          type: boolean
      required:
      - email
      - message
      - name
    ContactRequestResponseDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        message:
          type: string
        consentProcessing:
          type: boolean
        assignedToAccountId:
          type: string
          format: uuid
        assignedToName:
          type: string
        assignedToEmail:
          type: string
        assignedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    UpdateSupportConversationStatusRequestDto:
      type: object
      properties:
        status:
          type: string
          enum:
          - OPEN
          - CLOSED
      required:
      - status
...
    UpdateOrderStatusRequestDto:
      type: object
      properties:
        status:
          type: string
          enum:
          - CREATED
          - PROCESSING
          - PACKED
          - SHIPPED
          - DELIVERED
          - CANCELLED
        estimatedDeliveryAt:
          type: string
          format: date-time
        estimatedDeliveryEnd:
          type: string
          format: date-time
        receiptUrl:
          type: string
        comment:
          type: string
      required:
      - status
...
  securitySchemes:
    accessTokenCookie:
      type: apiKey
      description: |
        JWT access token (HttpOnly cookie). Выставляется после POST /auth/login.
        В Swagger UI: Authorize → значение токена (если копируете из DevTools) или вход через браузер на том же origin.
      name: GBCC_ACCESS_TOKEN
      in: cookie
    supportGuestToken:
      type: apiKey
      description: |
        Токен гостя для обращения в поддержку без аккаунта.
        Возвращается в теле ответа POST /support/conversations (поле guestAccessToken).
      name: X-Support-Token
      in: header